# docs/swagger.yaml
basePath: /
definitions:
  models.APINoteCreateSuccessResponse:
    properties:
      code:
        description: 状态码
        type: integer
      detail:
        allOf:
        - $ref: '#/definitions/models.NoteResponse'
        description: 详细信息
      msg:
        description: 状态信息
        type: string
    type: object
  models.APINoteDelSuccessResponse:
    properties:
      code:
        type: integer
      detail:
        type: string
      msg:
        type: string
    type: object
  models.APINoteFailed:
    properties:
      code:
        type: integer
      delete:
        type: string
      msg:
        type: string
    type: object
  models.APINotesGetSuccessResponse:
    properties:
      code:
        description: 状态码
        type: integer
      detail:
        description: 详细信息
        items:
          $ref: '#/definitions/models.NoteResponse'
        type: array
      msg:
        description: 状态信息
        type: string
    type: object
  models.APIResponse:
    properties:
      code:
        description: 状态码
        type: integer
      detail:
        description: 额外携带的详细信息
      msg:
        description: 状态消息
        type: string
    type: object
  models.CategoryName:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models.CreateNoteParam:
    properties:
      category_id:
        type: integer
      content_md:
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      title:
        type: string
    required:
    - category_id
    - content_md
    - title
    type: object
  models.LoginParam:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.NoteResponse:
    properties:
      author_name:
        type: string
      category_name:
        type: string
      contend_html:
        type: string
      contend_md:
        type: string
      create_at:
        type: string
      id:
        example: "0"
        type: string
      status:
        type: integer
      title:
        type: string
      update_at:
        type: string
    type: object
  models.RegisterParam:
    properties:
      email:
        type: string
      gender:
        maximum: 2
        minimum: 0
        type: integer
      password:
        type: string
      re_password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - re_password
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: 1044011439@qq.com
    name: Xu Mingjie
  description: 个人网站
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: 个人网站
  version: 1.0.0
paths:
  /api/v1/admin/categories:
    get:
      description: 获取所有分类名称
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取所有分类名称
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: 创建新的分类
      parameters:
      - description: 分类信息
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryName'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 添加一个分类
      tags:
      - categories
  /api/v1/admin/categories/{id}:
    delete:
      description: 根据ID删除分类
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除一个分类
      tags:
      - categories
    get:
      description: 根据ID查询特定分类
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 查询一个分类
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: 根据ID更新分类信息
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类名称
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/models.CategoryName'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新一个分类
      tags:
      - categories
  /api/v1/admin/notes:
    get:
      description: 获取所有笔记
      produces:
      - application/json
      responses:
        "200":
          description: 获取笔记成功
          schema:
            $ref: '#/definitions/models.APINotesGetSuccessResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.APINoteFailed'
      security:
      - ApiKeyAuth: []
      summary: 获取所有笔记
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: 创建新的笔记
      parameters:
      - description: 笔记信息
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/models.CreateNoteParam'
      produces:
      - application/json
      responses:
        "200":
          description: 创建笔记成功
          schema:
            $ref: '#/definitions/models.APINoteCreateSuccessResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.APINoteFailed'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.APINoteFailed'
      security:
      - ApiKeyAuth: []
      summary: 添加一个笔记
      tags:
      - notes
  /api/v1/admin/notes/{id}:
    delete:
      description: 根据ID删除笔记
      parameters:
      - description: 笔记ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除笔记成功
          schema:
            $ref: '#/definitions/models.APINoteDelSuccessResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.APINoteFailed'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/models.APINoteFailed'
      security:
      - ApiKeyAuth: []
      summary: 删除一个笔记
      tags:
      - notes
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 用户登录
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/models.LoginParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - user
  /user/signup:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 用户注册
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/models.RegisterParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 注册新用户
      tags:
      - user
swagger: "2.0"